// src/types/entities/securityData.spec.ts
export type TSecurityData = {
    CODGEO_2022: string | undefined;
    annee: string | undefined;
    classe: string | undefined;
    unité_de_compte: string | undefined;
    valeur_publiee: string | undefined;
    faits: string | undefined;
    tauxpourmille: string | undefined;
    complementinfoval: string | undefined;
    complementinfotaux: string | undefined;
    POP: string | undefined;
    millPOP: string | undefined;
    LOG: string | undefined;
    millLOG: string | undefined;
};